\name{callProc}
\alias{callProc}
\title{
  Call an SQL stored procedure
}
\description{
Calls a stored procedure on a remote RDBMS
}
\usage{
callProc(con, ...)
}
\arguments{
\item{con}{
a \code{dbConnection} object.
}
\item{\dots }{
additional arguments are passed to the implementing method.
}
}
\details{
Not yet implemented.
}
\keyword{interface}
\keyword{database}
% vim:syntax=tex
