% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\name{notice_msg}
\alias{notice_msg}
\title{Sample message}
\usage{
notice_msg(msg = paste(remoji::emoji("100"), "Done", remoji::emoji("+1")))
}
\arguments{
\item{msg}{oracle messsage by \code{R} (string)}
}
\description{
Sample message
}
\examples{
\dontrun{
notice_msg()
}
}

