% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notice_overtime.R
\name{notice_overtime}
\alias{notice_overtime}
\title{Notice running r code ends.}
\usage{
notice_overtime(expr, overtime = 0.1, msg = NULL)
}
\arguments{
\item{expr}{any syntactically valid \code{R} expression}

\item{overtime}{threshold}

\item{msg}{oracle messsage by \code{R} (string)}
}
\description{
Notice running r code ends.
}
\examples{
\dontrun{
notice_overtime(1 + 1) # None
notice_overtime(for(i in 1:1e7) sqrt(1:10)) # Showing messages.
}
}

