% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notice_status.R
\name{notice_status}
\alias{notice_status}
\title{Notice running r code evaluate status.}
\usage{
notice_status(expr = NULL, msg = NULL, timing = 2)
}
\arguments{
\item{expr}{any syntactically valid \code{R} expression}

\item{msg}{oracle messsage by \code{R} (string)}

\item{timing}{when display oracle messsage after evaluation.}
}
\description{
Notice running r code evaluate status.
}
\examples{
\dontrun{
notice_status(head(iris))

# With chunk statement
mtcars \%>\% dplyr::select(mpg, cyl, disp) \%>\%
dplyr::mutate(disp2 = disp / 2) \%>\%
  head() \%>\%
    notice_status()

plot(iris$Sepal.Length, iris$Petal.Width) \%>\%
  notice_status(msg = "DONE!")

library(ggplot2)
ggplot(iris, aes(Sepal.Length, Petal.Width)) +
geom_point() \%>\%
notice_status()
}
}

