% $Id: S4R.Rd st_server_demukhin_r/1 2011/07/22 22:11:35 vsashika Exp $
\name{S4R}
\alias{ErrorClass}

\title{R compatibility with S version 4/S-Plus 5+ support functions}
\description{
  These objects ease the task of porting functions into R
  from S Version 4 and S-Plus 5.0 and later.
  May be obsolete in the future.
}
\keyword{internal}
% vim:syntax=tex
