\name{summary-methods}
\docType{methods}
\alias{summary-methods}
\alias{show-methods}
\alias{show,OraDriver-method}
\alias{show,OraConnection-method}
\alias{show,OraResult-method}
\alias{summary,OraDriver-method}
\alias{summary,OraConnection-method}
\alias{summary,OraResult-method}
\title{Summarize an Oracle object}
\description{
   These methods are straight-forward implementations of the corresponding
   generic functions.
}
\keyword{methods}
\keyword{database}
\keyword{interface}
