\name{S4R}
\alias{ErrorClass}
\alias{AS}
\alias{NEW}
\alias{NEW.default}
\alias{NEW.OraObject}
\alias{NEW.OraManager}
\alias{NEW.OraConnection}
\alias{NEW.OraResultSet}
\alias{usingR}

\title{R compatibility with S version 4/Splus5+ support functions}
\description{
  These objects ease the task of porting functions into R
  from S Version 4 and S-Plus 5.0 and later.  See the documentation
  of the lower-case functions there. May be obsolete in the future.
}
\usage{
AS(object, classname)
NEW(classname, ...)
usingR(major, minor)
}
\keyword{internal}
% vim:syntax=tex
